requirements.txt
streamlit==1.32.0
python-dotenv==1.0.1
langgraph==0.0.24
langchain==0.1.12
langchain-core==0.1.32
langchain-groq==0.0.6
typing-extensions>=4.5.0
openai>=1.0.0